Covid Data for Ireland and the United Kingdom

covid_data_uk_irl returns a dataset that contains the wrangled covid19 related data for Ireland and the United Kingdom with the variables listed below.

covid_data_uk_irl

Format

covid_data_uk_irl returns a dataset that contains covid19 related data for Ireland and the United Kingdom, updated as of the 21st of October 2020.

date date in YYYY-MM-DD format

iso3 ISO3 country code (character)

Cumulative Confirmed CasesCumulative N reported COVID-19 cases up to and including this updated as of the 21st of October 2020.

Cumulative Confirmed Cases per capita, 100,000 Cumulative N reported COVID-19 cases up to and including this date, per capita 100,000 updated as of the 21st of October 2020.

Daily Death Total per capita, 100,000 N reported COVID-19 deaths updated as of the 21st of October 2020., per capita 100,000

Cumulative Death per capita, 100,000 Cumulative N reported COVID-19 deaths updated as of the 21st of October 2020.

Daily Confirmed Cases Number of daily reported COVID-19 cases on the chosen date up to the 21st of October 2020.

Cumulative DeathsCumulative N reported COVID-19 deaths up to the 21st of October 2020.

Daily Confirmed Deaths Number reported COVID-19 daily deaths on the chosen date up to the 21st of October 2020.

Daily Confirmed Cases per capita, 100,000N reported COVID-19 cases on this date, per capita 100,000

Log of Cumulative Confirmed Cases Log base 10 of cumulative N reported COVID-19 cases up to and including this date

Log of Daily Confirmed Cases Log base 10 of N reported COVID-19 cases on this date

Log of Cumulative Deaths Log base 10 of cumulative N reported COVID-19 deaths up to and including this date

Social Distancing measures Number of social distancing measures reported up to date by ACAPS, net of lifted restrictions

Movement Restrictions Number of movement restrictions reported up to date by ACAPS, net of lifted restrictions

Public Health Measures Number of public health measures reported up to date by ACAPS, net of lifted restrictions

details A list of information nested in the cell that contains the date and number of Social Distancing Measures, Movement Restrictions and Public Health Measures for that day

Source

The data has been collected from tidycovid19 package and the covdata package

Author

Samuel Lyubic

Examples

if (FALSE) {
# Directly call variables from the `covid_data_uk_irl` dataset
 covid_data_uk_irl %>%
select(`Log of Daily Confirmed Cases`, `Movement Restrictions`, `Public Health Measures`)
}