Generate a date range input

dateInput_range( ) will create a pair of text inputs from the the specified date variable which, when clicked on, bring up calanders that the user can click on to select the specific dates they wish to choose.

dateInput_range(
  inputId,
  label = "Chosen date range",
  date,
  start = min(date),
  ...
)

Arguments

inputId
The input slot that will be used to access the value.

label
Display label for the control, or NULL for no label.

date
The date input that contains the date information for the dataset. Either a Date object, or string in yyyy-mm-dd formart. Must be a date vector of at least length 2. A column from a dataset can be directly called using "$".

start
The initial start date. Either a Date object, or a string in yyyy-mm-dd format. If start is NULL, it will be reated as min(date) which is the default and the minimum date from the date variable provided will be selected.

...
The rest of the arguments that can be used have been sourced from the ?dateRangeInput and detailed below in "details":

Details

dateInput_range( ) will return a date range with chosen date values as specified by the user. The `date`` input can directly call a date variable from a dataset with the start, end and mximum dates being automatically generated from the provided date dataset if their fields are left NULL.

The date format string specifies how the date will be displayed in the browser. It allows the following values:

yy Year without century (12)

yyyy Year with century (2012)

mm Month number, with leading zero (01-12)

m Month number, without leading zero (1-12)

M Abbreviated month name

MM Full month name

dd Day of month with leading zero

d Day of month without leading zero

D Abbreviated weekday name

DD Full weekday name

The remaining arguments that can be used with dateInput_range( ) function are as follows: #

end The initial end date. Either a Date object, or a string in yyyy-mm-dd format. If end is NULL, the maximum date from the provided dataset will be chosen.

max The maximum allowed date. Either a Date object, or a string in yyyy-mm-dd format. If max is NULL, the maximum date from the provided dataset will be chosen.

startview The date range shown when the input object is first clicked. Can be "month" (the default), "year", or "decade".

weekstart Which day is the start of the week. Should be an integer from 0 (Sunday) to 6 (Saturday).

language The language used for month and day names. Default is "en". Other valid values include "ar", "az", "bg", "bs", "ca", "cs", "cy", "da", "de", "el", "en-AU", "en-GB", "eo", "es", "et", "eu", "fa", "fi", "fo", "fr-CH", "fr", "gl", "he", "hr", "hu", "hy", "id", "is", "it-CH", "it", "ja", "ka", "kh", "kk", "ko", "kr", "lt", "lv", "me", "mk", "mn", "ms", "nb", "nl-BE", "nl", "no", "pl", "pt-BR", "pt", "ro", "rs-latin", "rs", "ru", "sk", "sl", "sq", "sr-latin", "sr", "sv", "sw", "th", "tr", "uk", "vi", "zh-CN", and "zh-TW".

serperator String to display between the start and end input boxes.

width The width of the input, e.g. '400px', or '100%.

autoclose Whether or not to close the datepicker immediately when a date is selected. Takes arguments TRUE or FALSE.

format is the date format presented to the user, the options are detailed below. The default is yy-mm-dd.