Mobility data for the major cities in Ireland and the United Kingdom

mobility_major_cities returns a dataset that contains the mobility scores for the major cities in Ireland and the United Kingdom, as well as latitude and longitude location. All the variables are listed below.

mobility_major_cities

Format

mobility_major_cities returns a dataset that has 9,755 rows and 12 columns, as follows

geo_type character Type geographical unit. Values: city or country/region

region character Name of geographical unit.

transportation_typecharacter Mode of transport. Values: Driving, Transit, or Walking

alternative_name character Name of region in local language

sub_region character country Subregion

country character Country name (not provided for all countries)

date date Date in yyyy-mm-dddd format

scoredouble Activity score. Indexed to 100 on the first date of observation for a given mode of transport.

lat Latitude

lng Longitude

populationPopulation of the region

Details List of information containing city name and country

Source

The data has been collected from the covdata package

Details

Data from Apple Maps on relative changes in mobility in various cities and countries.

Author

Samuel Lyubic

Examples

if (FALSE) {
# Directly call variables from the `mobility_major_cities` dataset
mobility_major_cities %>%
ggplot(aes(x = lng, y = lat)) +
geom_point()
}