Raw form of the covid data used for analysis in the application

Covid data for Ireland and United Kingdom in the complete and raw form containing all the primary variables from the orginal collection.

raw_covid_data_uk_irl

Format

raw_covid_data_uk_irl returns a dataset that contains covid19 related data for Ireland and the United Kingdom, updated as of the 21st of October 2020.

date Date in YYYY-MM-DD format

cname Country name

iso3 ISO3 country code (character)

cu_casesCumulative number of reported COVID-19 cases up to and including this date updated as of the 21st of October 2020.

cu_cases_per_100k Cumulative N reported COVID-19 cases up to and including this date, per capita 100,000 updated as of the 21st of October 2020.

deaths_per_100k N reported COVID-19 deaths updated as of the 21st of October 2020., per capita 100,000

cu_deaths_per_100k Cumulative N reported COVID-19 deaths updated as of the 21st of October 2020.

cases_per_100k Daily confirmed reported COVID-19 cases on this date, per capita 100,000

log_cu_cases Log base 10 of cumulative N reported COVID-19 cases up to and including this date

log_cases Log base 10 of N reported COVID-19 cases on this date

log_cu_deaths Log base 10 of cumulative N reported COVID-19 deaths up to and including this date

cases Number of daily reported COVID-19 cases on the chosen date up to the 21st of October 2020.

cu_deathsCumulative N reported COVID-19 deaths up to the 21st of October 2020.

deaths Number reported COVID-19 daily deaths on the chosen date up to the 21st of October 2020.

recovered Covid-19 recoveries as reported by JHU CSSE (accumulated)

gcmr_retail_recreation Log base 10 of cumulative N reported COVID-19 cases up to and including this date

gcmr_grocery_pharmacy Log base 10 of N reported COVID-19 cases on this date

gcmr_transit_stations Log base 10 of cumulative N reported COVID-19 deaths up to and including this date

gcmr_workplaces Google Community Mobility Reports data for the frequency that people visit workplaces expressed as a percentage*100 change relative to the baseline period Jan 3 - Feb 6, 2020

Social Distancing measures Number of social distancing measures reported up to date by ACAPS, net of lifted restrictions

Movement Restrictions Number of movement restrictions reported up to date by ACAPS, net of lifted restrictions

Public Health Measures Number of public health measures reported up to date by ACAPS, net of lifted restrictions

pop population for each country in 2019

Source

The data has been collected from tidycovid19 package and the covdata package

Author

Samuel Lyubic

Examples

if (FALSE) {
# Directly call variables from the `raw_covid_data_uk_irl` dataset
raw_covid_data_uk_irl %>%
select(deaths, pop, cname)
}