Covid Data for Ireland and the United Kingdom

covid_data_uk_irl( ) returns a dataset that contains covid19 related data for Ireland and the United Kingdom

covid_data_uk_irl()

Format

covid_data_uk_irl( ) returns a dataset that contains covid19 related data for Ireland and the United Kingdom. A tibble with 494 rows and 18 columns, that consist of the following variables

date date in YYYY-MM-DD format

iso3 ISO3 country code (character)

Cumulative Confirmed CasesCumulative N reported COVID-19 cases up to and including this date

Cumulative Confirmed Cases per capita, 100,000 Cumulative N reported COVID-19 cases up to and including this date, per capita 100,000

Daily Death Total per capita, 100,000 N reported COVID-19 deaths on this date, per capita 100,000

Cumulative Death per capita, 100,000 Cumulative N reported COVID-19 deaths up to and including this date

Daily Confirmed Cases Cumulative N reported COVID-19 cases up to and including this date

Cumulative DeathsCumulative N reported COVID-19 deaths up to and including this date

Daily Confirmed Deaths N reported COVID-19 deaths on this date

Daily Confirmed Cases per capita, 100,000N reported COVID-19 cases on this date, per capita 100,000

Log of Cumulative Confirmed Cases Log base 10 of cumulative N reported COVID-19 cases up to and including this date

Log of Daily Confirmed Cases Log base 10 of N reported COVID-19 cases on this date

Log of Cumulative Deaths Log base 10 of cumulative N reported COVID-19 deaths up to and including this date

Social Distancing measures Number of social distancing measures reported up to date by ACAPS, net of lifted restrictions

Movement Restrictions Number of movement restrictions reported up to date by ACAPS, net of lifted restrictions

Public Health Measures Number of public health measures reported up to date by ACAPS, net of lifted restrictions

details A list of information nested in the cell that contains the date and number of Social Distancing Measures, Movement Restrictions and Public Health Measures for that day

Source

The data has been collected from tidycovid19 package and the covdata package

Author

Samuel Lyubic

Examples

if (FALSE) {
# To access the dataset as a whole 
covid_data_uk_irl()

# Create an object with the package and then call and work with any variable in the dataset
data <- covid_data_uk_irl()
data %>% select(`Log of Daily Confirmed Cases`, `Movement Restrictions`, `Public Health Measures`)
}